pocs/gemini-ai-vision-demo/pocs/pdf-item-extractor/requirements.txt
# PDF Item Extractor POC
# Agentic Vision 기반 PDF 문항 추출

# Google AI
google-genai>=1.0.0

# PDF 처리
PyMuPDF>=1.24.0

# 이미지 처리
Pillow>=10.0.0

# 데이터 검증
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 환경 변수
python-dotenv>=1.0.0
